feat: Move persistence drivers to own service and make them more generic
