+++ src/vue/forms/index.ts
@@ -1,9 +1,9 @@
 import { BaseForm, propertyAwareToRaw } from './BaseForm'
 import { type PersistedForm } from './types/PersistedForm'
-import { LocalStorageDriver } from './drivers/LocalStorageDriver'
-import { NonPersistentDriver } from './drivers/NonPersistentDriver'
-import { SessionStorageDriver } from './drivers/SessionStorageDriver'
-import { type PersistenceDriver } from './types/PersistenceDriver'
+import { LocalStorageDriver } from '../../service/persistenceDrivers/LocalStorageDriver'
+import { NonPersistentDriver } from '../../service/persistenceDrivers/NonPersistentDriver'
+import { SessionStorageDriver } from '../../service/persistenceDrivers/SessionStorageDriver'
+import { type PersistenceDriver } from '../../service/persistenceDrivers/types/PersistenceDriver'
 import { PropertyAwareArray, type PropertyAwareField, type PropertyAware } from './PropertyAwareArray'
 
 export { BaseForm, propertyAwareToRaw, PropertyAwareArray, NonPersistentDriver, SessionStorageDriver, LocalStorageDriver }
